[store] change: MetaNode::upsertKV return two equal value if no modification is applied
Previously, if an `Add` action failed, due to mismatching `seq`, it
returns `(previous value, None)`.
From now on it returns `(previous value, previous value)` to indicate
nothing change.

**Why**:

The previous solution can not distinguish the result from a successful
`delete` and the result from a failed `add`.
Thus the caller has to examine the context to determine if the return
value is a success. E.g., by considering the action(`add` or `delete`).

With the new solution, the caller knows if it succeeded by only a simple comparison on
`res.prev` and `res.result`.