[store]: feature: RaftState: stores the last snapshot meta. To avoid potential repliation bug.
Because membership log is not applied to state machine,
Before logs are removed, the membership need to be saved.

The async-raft::MemStore saves it in a removed log slot,
which may mess up the replication:
Replication has chance sending this pointer log that was an client data.

Although, the presence of snapshot meta does not mean there is a valid snapshot data.
Since snapshot data is not saved on disk.

A snapshot meta only guarantees no smaller snapshot will be installed
when installing snapshot, and no membership config is lost when removing
logs that are included in snapshot.