[ci] refactor: limit number of test running thread to 2.
Raft that is used in meta service is sensitive to time.
In a CI a unit test sometimes is delayed by one or two minutes and then fails due to timeout.

Such as this failure:
https://github.com/datafuselabs/databend/pull/1876/checks?check_run_id=3638769792

To reduce the max number of parallel test to run, may reduce the chance unit test delays each other.