[query] refactor: Table::read_plan does not need DatabendQueryContext anymore.
Why:
`Table` is a low level concept in databend code base and should be a
dependency of some other crate such as common/planner.

This commit is one of the steps to remove the `Table` dependency on `query`.

Trait `Table` references `DatabendQueryContext` as an argument type.
In this commit it is replaced with another smaller type `TableIOContext`
so that in future `Table` can be moved out of crate `query`.

`TableIOContext` provides data-access support, exposes the runtime used
by `query` itself and provides two resource value: max thread number and
node list.

- Add `TableIOContext` to provide everything a Table need to build a
  plan or else.

- `Table::read_plan()` use `TableIOContext` as argument to replace
  `DatabendQueryContext`.

- When calling `read_plan()`, a temporary `TableIOContext` is built out
  of DatabendQueryContext.

- `DatabendQueryContext` provides two additional supporting methods:
  `get_single_node_table_io_context()` and
  `get_cluster_table_io_context()`.

- fix: #2072