[query] refactor: Catalog: removed unused get_table_by_id()
After fetching a table from catalog by `db_name, table_name`,
the `TableInfo` should be kept and transferred between query nodes.

There should not be need to re-fetch a table by id, to execute a sql.
If there is, it may be a inappropriate design.

Thus `get_table_by_id()` is removed.
If there will be need to add it back,
we will reconsider it.