[common/meta/types] refactor: add TableMeta containing essential table info.
`table_name`, `table_id`, `table_version` and `desc` is not part of
what-a-table-is.
They are just about how to find a table etc.

Meta-store should just store what-a-table-is.
Other information should be maintained by upper layer logics.

By moving unnecessary info out of TableMeta, it prevents
inappropriate usage of these non-critical info, e.g.: trying to get a
table by name in a context in which table name is volatile.