[metasrv] feature: metasrv has to be compatible with 20220413-34e89c99
Let metasvr be able to load the log-entry format before commit
34e89c99e4e35632718e9227f6549b3090eb0fb9,
in which the `Cmd` format changed.

To solve th compatible issue, an intermedia type is introduced, which is
a super set of the old and new type.
When loading a record, first load it into the superset type, then reduce
to the latest type.

- fix: #4890