refactor(proto-conv): trait FromToPB use associated type instead of type parameter
Change trait signature `FromToProto<PB: prost::Message>` to

```rust
trait FromToProto {
    type PB;
}
```

When impl `FromToProto` for a rust type, there should be only one
protobuf type that is convert from and to it. I.e, the protobuf type is
determined by rust type.
Thus `PB` should be a associated type.

Otherwise, there could be two protobuf types bounded to one rust type(e.g., `impl
FromToProto<A> for T` and `impl FromToProto<B> for T`), which is
potentially invalid: when converting rust type to protobuf type, it does
not know which it should be converted to.