fix(meta): When handling append-entries, if prev_log_id is purged, it should not delete any logs
Update openraft to [v0.7.0-alpha.3](https://github.com/datafuselabs/openraft/tree/v0.7.0-alpha.3);

When handling append-entries, if the local log at `prev_log_id.index` is
purged, a follower should not believe it is a **conflict** and should
not delete all logs. It will get committed log lost.

To fix this issue, use `last_applied` instead of `committed`:
`last_applied` is always the committed log id, while `committed` is not
persisted and may be smaller than the actually applied, when a follower
is restarted.

- Detail: https://github.com/datafuselabs/openraft/issues/511

- Part of #7112