fix(meta): when expiring a record, leader and followers should use the same now time
When leader propose a raft-log, it embeds the local time in the log
entry.

When applying a log to state machine, use this time instead of clock
time to determine if a record is expired(if it has such a config).

This way the leader and the followers will generate the same state.