refactor(meta-service): clean up MetaError
MetaError sub errors:

Add `MetaStartupError` as a collection of errors raised when starting up
meta-service. These errors do not have to be part of MetaError.

Remove `MetaRaftError`. This error is not clear about when a error occurs.
Raft related errors could occur locally or on a remote peer.

Add `MetaAPIError`, which is the root error for a meta-service API.

`MetaNetworkError`: add InvalidArgument and InvalidReply.