refactor(meta/management): when joining a node, detect membership instead of log
A databend-meta node may always be started with `--join` arg.
With this arg, meta node should decide whether to execute the
**join-to-a-cluster** process.

Before this commit it detect if there are logs. This way is not very
relliable: it's possible the leader has setup a replication to this new
node but not yet added it as a **voter**. In such a case, this node will
never be added into the cluster automatically.

In stead, in this commit it detect if there is a **membership** config
that includes this node to decide if to join a cluster. Thus only when a
node has already joined to a cluster(leader committed the membership and
has replicated it to this node), it skips the join process.