refactor(meta): return InvalidReply if fail to decode a message
During upgrading, different version of databend-query/meta run in a same
cluster. If a message that can not be decoded received, return an
`InvalidReply` error instead of just panicking the program.

Returning such an error might make the cluster unable to operate while
upgrading, but data durability and consistency won't be compromised.