refactor(meta/client): add compatible layer for upgrade
Currently meta-service responds `KVAppError` for kv-api, which is too big.
`MetaAPIError`, which is a sub error of `KVAppError`,  would be enough.

This is the first step of this change: let the client be able to parse
both `KVAppError` and `MetaAPIError`. When this patch is merged, the
server side can upgrade to produce just `MetaAPIError`.

Changes on client side:

Add type `Compatible<Outer, Inner>` as a compatbile enum for both
`KVAppError` and `MetaAPIError`.

Try to parse a responded error into `Compatible` then extract
`MetaAPIError` for use.

It no longer uses `KVAppError`, e.g., SchemaApi or ShareApi just relies
on a KVApi implementation that just returns MetaError.

Changes on server side:

The meta sercie still responds `KVAppError`. But some of the error types
used internally are adjusted.