fix(meta): when starting up, it should not try to register until leader node to be sync-ed to local store
A node registers itself to the leader when starting up. Before the raft
log that contains leader node info is replicated to this node, it does
not know leader's address. The registering will fail and the process
will shutdown.

Therefore before registering, it has to wait for the raft log containing
leader node to be synced.