refactor: revert to using all endpoints if all are unhealthy.
Previously, the meta-grpc-client would eliminate the unhealthy endpoints
if the candidate endpoints contained more than one element.
However, it is still probable to obtain an empty list after filtering
out the unhealthy endpoints. The correct way should be to first
filter out the unhealthy endpoints and then verify if the list is empty.
If it is empty, it should revert to using all the endpoints,
regardless of their health status.

And the cached endpoints does not need an `async` lock.

- Part of: #10829