feat(meta-service): Split log store and state-machine store (#11124)
* chore: minor refactor

* feat(meta-service): Split log store and state-machine store

Separate the log store and state-machine store, in order to prevent
blocking of append-entries (heartbeat) when creating a snapshot.

[Openraft](https://github.com/datafuselabs/openraft/tree/release-0.8) changes since v0.8.4-alpha.1 to v0.8.4-alpha.2 :

### Feature: add feature flag storage-v2 to enable RaftLogStorage and RaftStateMachine

`storage-v2`: enables `RaftLogStorage` and `RaftStateMachine` as the v2 storage
This is a temporary feature flag, and will be removed in the future, when v2 storage is stable.
This feature disables `Adapter`, which is for v1 storage to be used as v2.
V2 storage separates log store and state machine store so that log IO and state machine IO can be parallelized naturally.

### Improve: getting a snapshot does not block RaftCore task

`RaftCore` no longer blocks on receiving a snapshot from the state-machine
worker while replicating a snapshot. Instead, it sends the `Receiver` to
the replication task and the replication task blocks on receiving the
snapshot.

### Fix: compat07::SnapshotMeta should decode v08 SnapshotMeta

### Fix: if the application does not persist snapshot, build a snapshot when starting up

### Improve: move state machine operations to another task

State machine operations, such as applying log entries, building/installing/getting snapshot are moved to `core::sm::Worker`, which is run in a standalone task other than the one running `RaftCore`.
In this way, log io operation(mostly appending log entries) and state machine io operations(mostly applying log entries) can be paralleled.

- Log io are sitll running in `RaftCore` task.

- Snapshot receiving/streaming are removed from `RaftCore`.

- Add `IOState` to `RaftState` to track the applied log id.

  This field is used to determine whether a certain command, such as
  sending a response, can be executed after a specific log has been
  applied.

- Refactor: `leader_step_down()` can only be run when the response of the second change-membership is sent.
  Before this commit, updating the `committed` is done atomically with
  sending back response. Since thie commit, these two steps are done
  separately, because applying log entries are moved to another task.
  Therefore `leader_step_down()` must wait for these two steps to be
  finished.

### Change: move `RaftStateMachine` out of `RaftStorage`

In Raft, the state machine is an independent storage component that
operates separately from the log store. As a result, accessing the log
store and accessing the state machine can be naturally parallelized.

This commit replaces the type parameter `RaftStorage` in
`Raft<.., S: RaftStorage>` with two type parameters: `RaftLogStorage` and
`RaftStateMachine`.

- Add: `RaftLogReaderExt` to provide additional log access methods based
  on a `RaftLogReader` implementation. Some of the methods are moved
  from `StorageHelper` to this trait.

- Add: `Adapter` to let application use the seperated log state machine
  framework without rewriting `RaftStorage` implementation.

### Fix: ProgressEntry::is_log_range_inflight() checks a log range, not a log entry This bug causes replication tries to send pruged log.

### Change: remove defensive check utilities

Most defensive checks are replaced with `debug_assert!` embedded in Engine.
`StoreExt` as a `RaftStorage` wrapper that implements defensive checks
are no longer needed. `StoreExt` are mainly used for testing and it is
very slow so that can not be used in production.

- Remove structs: `StoreExt`, `DefensiveStoreBuilder`
- Remove traits: `Wrapper`, `DefensiveCheckBase`, `DefensiveCheck`,

### Change: remove unused trait RaftStorageDebug

`RaftStorageDebug` has only one method `get_state_machine()`,
and state machine is entirely a user defined struct. Obtaining a state
machine does not imply anything about the struct or behavior of it.

### Refactor: allow snapshot buillding and streaming at the same time

Although enabling snapshot building and streaming at the same time may
lead to inefficiencies, it enhances the clarity of the system's overall
structure. For instance, when processing a stream, there is
no need to handle any building state.

In this commit, a new `snapshot_state::State` is introduced that includes
both a building state and a streaming state.

### Improve: send AppendEntries response before committing entries

When a follower receives an append-entries request that includes a
series of log entries to append and the log id that the leader
has committed, it responds with an append-entries response after
committing and applying the entries.

However, this is not strictly necessary. The follower could simply send
the response as soon as the log entries have been appended and flushed
to disk, without waiting for them to be committed.

### Improve: follower stores the accepted log id

A follower stores the ID of the latest log that has been synchronized by
the leader, represented as `(LeaderId, Option<LogId>)`.
This information is only kept in memory and is not taken into account
when processing vote requests or determining the commit condition.

The reason for storing this information is to allow followers to commit
only those logs that have been replicated by their current leader.

* refactor(meta-service): test building snapshot not block append-entries

Add a config for debugging that temporarily blocks snapshot building,
for testing purposes.

Add two tests that verify the leader's ability to write log entries
while a follower's snapshot building is blocked, in a 2-voter cluster.