feat(meta-service): write data version to local store
Write a data header to the raft-store so that compatibility can be
asserted and dealt with. When starting up, if there is no data version,
a default one `V0` will be inserted.

The default version is `V0` which is also the working data version in
this commit.

`metactl --export` always export the data version in a `Header` record
at the first line.

`metactl --import` will check the data version at the first line.
If the version is present and is not `V0`, it will quit at once with an
error.

- Fix: #4950