fix(meta): upgrade openraft to 0.8.4-alpha.3; bug fix, improve performance, etc
# Changes Summary:

- Fix: replication should be able to shutdown when replicating snapshot to unreachable node.
- Feature: new RaftNetwork API with argument RCPOption.
- Feature: add `RaftMetrics.vote`, `Wait::vote()`.
- Feature: add Wait::purged() to wait for purged to become the expected.
- Feature: add RaftMetrics::purged to report the last purged log id.
- Feature: add backoff strategy for unreachable nodes.
- Improve: build snapshot in anohter task.
- Improve: Better performance with two channel: `256 clients, 3 node: 1 million rps`.

# Detailed change log

## Fix: replication should be able to shutdown when replicating snapshot to unreachable node

If a replication is sending a snapshot, it should
periodically verify the input channel's status. When the input channel
is closed during replication rebuilding, it should immediately exit the
loop instead of attempting retries indefinitely.

## Change: move snapshot type definition from storage traits to `RaftTypeConfig`

Similar to `NodeId` or `Entry`, `SnapshotData` is also a data type that
is specified by the application and needs to be defined in
`RaftTypeConfig`, which is a collection of all application types.

Public types changes:

- Add `SnapshotData` to `RaftTypeConfig`:
  ```rust
  pub trait RaftTypeConfig {
      /// ...
      type SnapshotData: AsyncRead + AsyncWrite + AsyncSeek + Send + Sync + Unpin + 'static;
  }
  ```
- Remove associated type `SnapshotData` from `storage::RaftStorage`.
- Remove associated type `SnapshotData` from `storage::v2::RaftStateMachine`.

Corresponding API changes:

- Change `storage::RaftSnapshotBuilder<C: RaftTypeConfig, SNAPSHOT_DATA>` to `RaftSnapshotBuilder<C>`
- Change `storage::Snapshot<NID: NodeId, N: Node, SNAPSHOT_DATA>` to `storage::Snapshot<C>`

## Change: `RaftMetrics.replication` type to `BTreeMap<NodeId, Option<LogId>>`

The `RaftMetrics.replication` used to be of type
`ReplicationMetrics{ replication: BTreeMap<NodeId, ReplicationTargetMetrics> }`
which contained an atomic log index value for each
ReplicationTargetMetrics stored in the `BTreeMap`. The purpose of this
type was to reduce the cost of copying a metrics instance. However,
since the metrics report rate has been significantly reduced, this
cost is now negligible. As a result, these complicated implementations
have been removed. When reporting metrics, they can simply be cloned
from the progress information maintained by `Engine`.

## Feature: `RaftNetwork::send_append_entries()` can return PartialSuccess

If there are too many log entries and the `RPCOption.ttl` is not
sufficient, an application can opt to only send a portion of the
entries, with `AppendEntriesResponse::PartialSuccess(Option<LogId>)`, to
inform Openraft with the last replicated log id. Thus replication can
still make progress.

For example, it tries to send log entries `[1-2..3-10]`, the application
is allowed to send just `[1-2..1-3]` and return `PartialSuccess(1-3)`,

### Caution

The returned matching log id must be **greater than or equal to** the
first log id(`AppendEntriesRequest::prev_log_id`) of the entries to
send. If no RPC reply is received, `RaftNetwork::send_append_entries`
**must** return an `RPCError` to inform Openraft that the first log
id(`AppendEntriesRequest::prev_log_id`) may not match on the remote
target node.

## Feature: new RaftNetwork API with argument RCPOption

- `RaftNetwork` introduced 3 new API `append_entries`,
  `install_snapshot` and `vote` which accept an additional argument
  `RPCOption`, and deprecated the old API `send_append_entries`,
  `send_install_snapshot` and `send_vote`.

- The old API will be **removed** in `0.9`. An application can still
  implement the old API without any changes. Openraft calls only the new
  API and the default implementation will delegate to the old API.

- Implementing the new APIs will disable the old APIs.

- The new APIs accepts an additional argument `RPCOption`, to enable an
  application control the networking behaviors based on the parameters
  in `RPCOption`.

  The `hard_ttl()` and `soft_ttl()` in `RPCOption` sets the hard limit
  and the moderate limit of the duration for which an RPC should run.
  Once the `soft_ttl()` ends, the RPC implementation should start to
  gracefully cancel the RPC, and once the `hard_ttl()` ends, Openraft
  will terminate the ongoing RPC at once.

## Feature: add `RaftMetrics.vote`, `Wait::vote()`

The latest approved value of `Vote`, which has been saved to disk, is
referred to as `RaftMetrics.vote`. Additionally, a new `vote()` method
has been included in `Wait` to enable the application to wait for `vote`
to reach the anticipated value.

## Feature: add Wait::purged() to wait for purged to become the expected

## Feature: add RaftMetrics::purged to report the last purged log id

## Feature: add backoff strategy for unreachable nodes

Implements a backoff strategy for temporarily or permanently unreachable nodes.
If the `Network` implementation returns `Unreachable` error, Openraft
will backoff for a while before sending next RPC to this target.
This mechanism prevents error logging flood.

Adds a new method `backoff()` to `RaftNetwork` to let an application
return a customized backoff policy, the default provided backoff just
constantly sleep 500ms.

Adds an `unreachable_nodes` setting to the testing router `TypedRaftRouteryped` to emulate unreachable nodes.
Add new error `Unreachable` and an `RPCError` variant `Unreachable`.

## Improve: build snapshot in anohter task

Before this commit, snapshot is built in the `sm::Worker`, which blocks
other state-machine writes, such as applying log entries.

This commit parallels applying log entries and building snapshot: A
snapshot is built in another `tokio::task`.

Because building snapshot is a read operation, it does not have to
block the entire state machine. Instead, it only needs a consistent view
of the state machine or holding a lock of the state machine.

## Improve: move receiving snapshot chunk to sm::Worker

Receiving snapshot chunk should not be run in RaftCore task.
Otherwise it will block RaftCore.

In this commit this task is moved to sm::Worker, running in another
task. The corresponding responding command will not be run until
sm::Worker notify RaftCore receiving is finished.