refactor(meta): persist snapshot instead of state-machine
Make state machine a pure in-memory store and the durability relies on
persistent snapshot.

Upgrade to openraft v0.8.4-alpha.8: with the ability to persist
`committed` log id, so that upon startup, state machine will be able to
replay all committed log to restore to the latest state.

### Compatibility

This commit add a new data version V002, which depends on upgrading upon
startup to provide compatibility.

New binary can run directly with old version data, `V0` or `V001`.
**Rolling upgrade** is allowed.

`databend-metactl` is also upgraded that can import `V0`, `V001` and
`V002` data, and export `V002` data. Old version data will be upgraded
immediately once imported.

#### Incompatible snapshot format

Snapshot format changed thus during rolling upgrading, it requires all
node data to be up to date, ensure there is no need to replicate with
snapshot.

### New State machine

The new state machine is a multi level in-memory map: only the top level
accepts modifications, the lower levels are read-only. This way building
snapshot won't block writing.

### Snapshot

When a new snapshot is built, it is persisted on disk. And the last 3
snapshot will be kept, older snapshot will be removed.

Snapshot file contains lines of json string. The first line is a
data-version header(also a json) to identify the verson of the snapshot.
The snapshot version is the same as the program data-version, e.g.,
`V002`.

### Other changes

- meta-cluster-status(/v1/cluster/status): add 2 fields: snapshot_last_log_id and purged