Refactor: eliminate ambiguous cache API (#12519)
In the `Cache` trait, the `get_mut()` and `peek_mut()` methods return
mutable references that permit users to modify the stored value.
However, when the updated value is given back into the cache, the
eviction policy determined by `Meter` is not validated against these
changes. This could potentially lead to risks, such as the cache
surpassing its maximum capacity.

Given that these two APIs are currently unused, my suggestion is to
remove them to prevent potential misuse in the future.