fix: avoid relying on db-id-to-name mapping when dropping a table
When dropping a table, it needs to update the count of all tables
belonging to a `tenant`.

Before this commit, it loads the `tenent` from a reversed mapping from
`database-id` to `tenant`.

In this commit, `tenent` is passed in to `schema-api` by callers.
So that when the reversed mapping is lacking(old data), it won't return
an UnknownDatabaseId error.