feat: meta-client supports both new stream api kv_read_v1()(since 1.2.163) and old api kv_api()(since 1.1.32)
When a connection is made to a meta-server, it returns a server-version
to the caller.

If server-version is 1.2.63 or newer, list-kvs uses the stream api
`kv_read_v1()`, otherwise, it reverts to old `kv_api()`.

A return type of the Stream API has been alerted to `BoxStream`. This is
constructed from an iterator when it reverts to using `kv_api()`.

This committee does not change the compatibility of databend-query. It
remains compatible with 1.1.32 and onwards.