refactor: state machine API MapApi::range() returns static lifetime stream
When querying the meta-service state machine with a stream oriented API,
such as `list_kv()`, the gRPC server respond to the client with a
`BoxStream<'static, ...>`. Therefore state machine has to return a
`'static` stream too.

In this commit, A new Stream `EntryStream<K>` is introduced as a type of
the output type for `MapApi::range()`