fix: address the slow range() function in the state machine
A State Machine is a structure with multiple levels, with the top level
being writable.

The top writable level cannot be borrowed to create a static stream.
Thus the range result must be copied.

In this commit, we minimize the top level as much as possible.
Additionally, when querying a range from the static shared label, we avoid copying the result.