refactor: move to meta-service handshake into poll
Before this commit the connection pool stores raw connections `Channel`.
Before using one, the client handshake with the server everytime using a
channel.

In this commit, the handshake is moved into the connection pool and the
pool stores `MetaServiceClient` that already finished the handshake
phase.

Thus the client no longer needs to send a handshake RPC before using an
in-pool connection.

Other changes:

- Add meta gRPC client re-connect test
- Add: EstablishedClient, when handshake expire, drop the client and create a new one.