refactor: optimize `SchemaApi::create_table()`
In this commit, when `create_table()` encounters a transaction conflict,
it re-fetch the latest state for next retry at once in the transaction
response.

This way, in each retry loop, it saves a few RPC call for fetching the
states. There is only one RPC in each loop, which is sending the transaction.

Because the state is fetched at once when transaction conflict, backoff
sleep time is removed between each retry, because sleeping just
increases the chance that the fetched state becoming stale.

Implementation of this refactor includes:
- Fetch required state before retry-loop.
- In each loop, validate the state and then send transaction with
  assertions and operations, just as before.
- If it conflicts, the latest state is returned at once for next retry.

Other changes:
- Add two small backoff 2ms and 5ms

- Fix: #14340