refactor: tenant is not allowed to be empty
In this commit, when loading databend-query config, if `tenant` is empty
string, return an error and shutdown.
This way, the in-process `tenant` is always non-empty and is stored in a
`NonEmptyString`, and it get rid of all empty-tenant error handling.

The return value of `get_tenant()` is also changed to `NonEmptyString`.