feat(query): check view's inner table privilege when create or alter view (#14986)
* feat: check view's inner table privilege when create or alter view

Note:

Select view no need to check the inner table privilege.

Because the user only can query view when the admin user grant select on
the view to the user.

So the admin user needs to know why the user need access this view.

This pr is avoid that:

User only has create or alter on db. And then the user can query table
as create or alter view. Like this:

```sql
--root
grant create on default.* to a;
--user a
create view v_t1 as select * from t1; -- If it can be success, the user
a can query table t1 according to query view v_t1. It maybe dangerous.
```

* create view only need check the user has select privilege on inner tables

* add test: create view as select view