refactor: introduce `UninitTenant` and `TIdentRaw`
`Tenant` is a meta-service key component representing a tenant,
with a embedded per-tenant config.

`UninitTenant` is a shadow of `TIdent` but without config thus it can
not be used as a key.
`UninitTenant` can be initialized to build a `Tenant` for use as a
meta-service key.

Correspondingly, this commit introduce `TIdentRaw` as an uninitialized
version of `TIdent`.
`TIdentRaw` does not cantain per-tenant config and is only used inside
meta-service values(not keys), which could be `serde`.

Other changes:

Populate usage of `Tennat` to `Catalog` implementation.