fix: Respond with `SnapshotMismatch` error instead of internal error
This fix updates OpenRaft to version 0.9.8 to address an issue where,
upon receiving a `SnapshotMismatch` error, the snapshot should be resent
from the beginning. This ensures that the leader can correctly reset and
reinitiate the snapshot transmission process.

And meta-service should return `SnapshotMismatch` error instead of
returning an internal gRPC error.