fix: Prevent premature snapshot commit on unfinished writes (#15534)
Prior to this commit, the snapshot writer would attempt to commit the
snapshot upon closure of the input channel, operating under the
assumption that the snapshot was fully written. This assumption is
flawed as the closure of the channel might be due to a process shutdown,
not completion of the snapshot.

To address this issue, this commit introduces a change where items sent
to the snapshot writer are encapsulated within a `WriteEntry`, which can
be either `Data(T)` or `Commit`. The snapshot is only considered
complete and ready for commit when a `Commit` variant is received. This
adjustment ensures that premature snapshot commits are avoided in cases
of early channel closures.