refactor: improve raft connection and logging filter
Only reuse a connection if there is no error returned from the last RPC.
When a raft RPC is finished, push the connection back for next RPC if it
returns an `Ok`. Otherwise it drops the connection and the next RPC will
re-create a new one. Such approach ensures no problematic connection
will be used.

Support multi log prefix filter.