feat: transfer leader command for meta-service
This commit introduce a new admin API(at `admin_api_address`, by default
`127.0.0.1:28002`) to force the current Leader to transfer leadership to
another node.  The url of this command is
"/v1/ctrl/trigger_transfer_leader?to=<new_leader_node_id>".

Such request returns nothing no matter leadership is successfully
transferred or not.

If the node receives this request is not the leader it does nothing.