refactor: introduce new method `accurate_columns_ranges` to `Table` (#16247)
* refactor: introduce new method `accurate_columns_ranges` to `Table`

Introduce a new method `accurate_columns_ranges` to the `Table` trait,
which returns the "accurate" ranges of the given columns.

The bounds of the column ranges are derived from the segment information
of the given table, making them tight. However, for string-type columns,
the bounds may have been truncated, which is indicated by the
`may_be_truncated` field in the `Bound` struct.

* tweak signature

* add unit test