refactor: add associated type kvapi::Value::KeyType
This way the key tyep and the value type becomes one-to-one mapping.
The KeyType is used to build a key from a value.

For example, `Id<DatabaseId>` is a value, to build a key to load
`DatabaseMeta`, it needs a `Tenant` to build the key `TIdent<T>`.
Such a `Tenant` can be retrieve from the key of `Id<DatabaseId>`, i.e.,
`<Id<DatabaseId> as kvapi::Value>::KeyType`

refactor: simplify get_data_mask