fix: handle trailing zeros in Raft log WAL with EXT4 writeback mode
When EXT4 is mounted with data=writeback, data and metadata (file length)
can be written to disk in arbitrary order, potentially leaving trailing
zeros in the WAL tail.

Fix: Truncate zero bytes starting from the first un-decodable WALRecord
and treat the chunk as successfully opened.

This issue is fixed in the dependent crate:
- https://github.com/drmingdrmer/raft-log/commit/757542bfc6f862184ee6f40a2d98c8222653767f